+++ src/main/java/br/com/greenpayimpact/calculadora/controller/CalculoController.java
@@ -1,16 +1,17 @@
 package br.com.greenpayimpact.calculadora.controller;
 
-import br.com.greenpayimpact.calculadora.dto.CalculoRequest;
-import br.com.greenpayimpact.calculadora.dto.CalculoResponse;
-import br.com.greenpayimpact.calculadora.model.Empresa;
-import br.com.greenpayimpact.calculadora.service.EmpresaService;
-import jakarta.validation.Valid;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
-import java.util.Map;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.greenpayimpact.calculadora.dto.CalculoResponse;
+import br.com.greenpayimpact.calculadora.service.EmpresaService;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -20,18 +21,7 @@ public class CalculoController {
     @Autowired
     private EmpresaService empresaService;
 
-    @PostMapping("/empresas")
-    public ResponseEntity<?> cadastrarEmpresa(@RequestBody @Valid CalculoRequest request, BindingResult result) {
-        if (result.hasErrors()) {
-            String mensagemErro = result.getAllErrors().get(0).getDefaultMessage();
-            return ResponseEntity.badRequest().body(Map.of("erro", mensagemErro));
-        }
-
-        Empresa empresaSalva = empresaService.salvarEmpresa(request);
-        return ResponseEntity.ok(empresaSalva);
-    }
-
-    @GetMapping("/calcular-impacto/{id}")
+    @GetMapping("/empresas/{id}/impacto")
     public ResponseEntity<?> calcularImpacto(@PathVariable Long id) {
         try {
             CalculoResponse response = empresaService.calcularImpactoPorId(id);
@@ -39,9 +29,5 @@ public class CalculoController {
         } catch (RuntimeException e) {
             return ResponseEntity.badRequest().body(Map.of("erro", e.getMessage()));
         }
-    }
-    @GetMapping("/empresas")
-    public ResponseEntity<List<Empresa>> listarEmpresas() {
-        return ResponseEntity.ok(empresaService.listarTodas());
     }
 }
