+++ src/main/java/br/com/greenpayimpact/calculadora/controller/FatorEmissaoController.java
@@ -1,14 +1,21 @@
 package br.com.greenpayimpact.calculadora.controller;
 
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import br.com.greenpayimpact.calculadora.dto.FatorEmissaoRequest;
 import br.com.greenpayimpact.calculadora.dto.FatorEmissaoResponse;
 import br.com.greenpayimpact.calculadora.service.FatorEmissaoService;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -19,10 +26,14 @@ public class FatorEmissaoController {
     private FatorEmissaoService service;
 
     @PostMapping
-    public ResponseEntity<FatorEmissaoResponse> cadastrarFator(@RequestBody @Valid FatorEmissaoRequest request) {
+    public ResponseEntity<?> cadastrarFator(@RequestBody @Valid FatorEmissaoRequest request) {
+    try {
         FatorEmissaoResponse response = service.adicionarFator(request);
         return ResponseEntity.ok(response);
+    } catch (RuntimeException e) {
+        return ResponseEntity.badRequest().body(Map.of("erro", e.getMessage()));
     }
+}
 
     @GetMapping
     public ResponseEntity<List<FatorEmissaoResponse>> listarHistorico() {
